Factor out core components into bootstrap-chat package
