+++ packages/bootstrap-chat/.gitattributes
@@ -1,0 +1,1 @@
+build/bundle.js binary
