+++ packages/bootstrap-chat/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "bootstrap-chat",
+  "version": "1.0.0",
+  "description": "",
+  "main": "build/bundle.js",
+  "scripts": {
+    "dev": "webpack --mode development -q",
+    "dev:watch": "webpack --mode development -q --watch"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "@typescript-eslint/eslint-plugin": "2.x",
+    "@typescript-eslint/parser": "2.x",
+    "babel-loader": "^8.1.0",
+    "css-loader": "^3.5.1",
+    "eslint": "6.x",
+    "eslint-config-react-app": "^5.2.1",
+    "eslint-plugin-flowtype": "4.x",
+    "eslint-plugin-import": "2.x",
+    "eslint-plugin-jsx-a11y": "6.x",
+    "eslint-plugin-react": "7.x",
+    "eslint-plugin-react-hooks": "2.x",
+    "mephisto-task": "^1.0.10",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "style-loader": "^1.1.3",
+    "url-loader": "^4.1.0",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11"
+  },
+  "peerDependencies": {
+    "mephisto-task": "^1.0.10",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  },
+  "dependencies": {
+    "axios": "^0.19.2",
+    "babel-eslint": "10.x",
+    "bootstrap": "4.3.1",
+    "bowser": "^2.9.0",
+    "jquery": "3.5.0",
+    "rc-slider": "^9.3.1",
+    "react-bootstrap": "0.32.4"
+  }
+}
