+++ packages/bootstrap-chat/src/ConnectionIndicator.jsx
