+++ packages/bootstrap-chat/src/Glyphicon.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Glyphicon({ name }) {
+  return (
+    <div className="icon-container">
+      <span className={`glyphicon glyphicon-${name}`} aria-hidden="true" />
+    </div>
+  );
+}
+
+export default Glyphicon;
