+++ packages/bootstrap-chat/src/TextResponse.jsx
