+++ packages/bootstrap-chat/src/VolumeControl.jsx
