+++ packages/bootstrap-chat/src/WorkerChatPopup.jsx
