+++ packages/bootstrap-chat/src/index.js
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2017-present, Facebook, Inc.
+ * All rights reserved.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+import ConnectionIndicator from "./ConnectionIndicator.jsx";
+import ReviewButtons from "./ReviewButtons.jsx";
+import TextResponse from "./TextResponse.jsx";
+import FormResponse from "./FormResponse.jsx";
+import VolumeControl from "./VolumeControl.jsx";
+import WorkerChatPopup from "./WorkerChatPopup.jsx";
+import SystemMessage from "./SystemMessage.jsx";
+import Glyphicon from "./Glyphicon.jsx";
+import ChatMessage from "./ChatMessage.jsx";
+import DoneButton from "./DoneButton.jsx";
+import DoneResponse from "./DoneResponse.jsx";
+import ConnectionStatusBoundary from "./ConnectionStatusBoundary.jsx";
+import DefaultTaskDescription from "./DefaultTaskDescription.jsx";
+
+export {
+  ConnectionIndicator,
+  ReviewButtons,
+  TextResponse,
+  FormResponse,
+  VolumeControl,
+  WorkerChatPopup,
+  Glyphicon,
+  SystemMessage,
+  DoneResponse,
+  DoneButton,
+  ChatMessage,
+  ConnectionStatusBoundary,
+  DefaultTaskDescription,
+};
